Split Interpolation and Tempo test files
